Add possibility to define custom map feature stroke color
